+++ plugins/kiali/src/pages/Kiali/KialiNoAnnotation.tsx
@@ -6,7 +6,7 @@ import {
   Page,
 } from '@backstage/core-components';
 
-import { ANNOTATION_SUPPORTED } from '../../Router';
+import { ANNOTATION_SUPPORTED } from '../../components/Router';
 
 export const KialiNoAnnotation = () => {
   return (