+++ plugins/kiali/tests/kiali.spec.ts
@@ -1,0 +1,23 @@
+import { expect, Page, test } from '@playwright/test';
+
+test.describe('Kiali plugin', () => {
+  let page: Page;
+  test.describe('kiali errors', () => {
+    test.beforeAll(async ({ browser }) => {
+      const context = await browser.newContext();
+      page = await context.newPage();
+      await page.goto('/kiali-error');
+      await page.locator('[data-test="Kiali Errors"]');
+    });
+
+    test.afterAll(async ({ browser }) => {
+      await browser.close();
+    });
+
+    test('Networking error', async () => {
+      await expect(
+        page.locator('[data-test="Warning: Error reaching Kiali"]'),
+      ).toBeDefined();
+    });
+  });
+});