+++ plugins/acr/CHANGELOG.md
@@ -1,3 +1,11 @@
+## @janus-idp/backstage-plugin-acr [1.3.5](https://github.com/janus-idp/backstage-plugins/compare/@janus-idp/backstage-plugin-acr@1.3.4...@janus-idp/backstage-plugin-acr@1.3.5) (2024-05-02)
+
+
+
+### Dependencies
+
+* **@janus-idp/cli:** upgraded to 1.8.5
+
 ## @janus-idp/backstage-plugin-acr [1.3.4](https://github.com/janus-idp/backstage-plugins/compare/@janus-idp/backstage-plugin-acr@1.3.3...@janus-idp/backstage-plugin-acr@1.3.4) (2024-05-02)