+++ plugins/kiali-backend/CHANGELOG.md
@@ -1,3 +1,11 @@
+## @janus-idp/backstage-plugin-kiali-backend [1.10.11](https://github.com/janus-idp/backstage-plugins/compare/@janus-idp/backstage-plugin-kiali-backend@1.10.10...@janus-idp/backstage-plugin-kiali-backend@1.10.11) (2024-05-02)
+
+
+
+### Dependencies
+
+* **@janus-idp/cli:** upgraded to 1.8.5
+
 ## @janus-idp/backstage-plugin-kiali-backend [1.10.10](https://github.com/janus-idp/backstage-plugins/compare/@janus-idp/backstage-plugin-kiali-backend@1.10.9...@janus-idp/backstage-plugin-kiali-backend@1.10.10) (2024-05-02)